Initial strict mode implementation
Currently turns on some existing checks for all versions of quickbook,
and makes it an error if a section isn't closed at the end of a
document. Probably need to spend more time combing through the code for
more places it should be stricter.

Also, added a nasty hack to line-compare, because I couldn't work out
how to stop Boost Build from passing the testing parameters to
line-compare. So now it ignores --strict if passed at the command line.